requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
pydantic==2.9.2
passlib==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
PyJWT==2.9.0
